==> nb/1.hetionet_computation/2.compute_calibration_metrics.ipynb <==
import pathlib

import numpy as np
import pandas as pd
import sklearn.metrics
import tqdm
import xswap

def dataframe_to_summary_metrics_old(df, metaedge):
    nrow = len(df)
    
    # Compute features and hold in memory as numpy arrays (less memory use than pandas column)
    degree_product = df['source_degree'].values * df['target_degree'].values
    scaled_degree = (degree_product / degree_product.max()).astype(np.float32)
    del degree_product
    
    analytic_prior = xswap.prior.approximate_xswap_prior(df['source_degree'].values,
                                                         df['target_degree'].values,
                                                         df['edge'].sum())
    del df['source_degree'], df['target_degree']
    
    metrics = list()
    for feature in ['xswap_prior', 'scaled_degree', 'analytic_prior']:        
        if feature == 'analytic_prior': 
            df['analytic_prior'] = analytic_prior
            del analytic_prior
        elif feature == 'scaled_degree':
            df['scaled_degree'] = scaled_degree
            del scaled_degree
            
        # Compute fraction and number of duplicates for each feature value
        row = (
            df
            .groupby(feature)
            .agg({'edge': ['count', 'mean']})
            .reset_index()
        )
        
        # Scored edges for later AUROC computation. (So df[feature] can be deleted ASAP)
        y_score = df[feature].values
        del df[feature]
        
        # Reformat from Dict[Tuple[str, str]: pd.Series] to Dict[str: np.ndarray]
        row = {
            feature: row[(feature, '')].values,
            'count': row[('edge', 'count')].values.astype(np.uint32),
            'fraction_edges': row[('edge', 'mean')].values.astype(np.float32),
        }
        
        # Save test edges
        y_true = df['edge'].values.astype(np.ushort)
        
        # Summarize feature values into single statistics
        row = {
            'feature': feature,
            'metaedge': metaedge,
            'cal': (row['count'] * (row[feature] - row['fraction_edges']) ** 2).sum() / nrow,
            'weighted_mae': np.average(np.abs(y_true - y_score), 
                                       weights=((y_true / y_true.sum()) + ((1 - y_true) / (1 - y_true).sum()))),
            'ref': (row['count'] * row['fraction_edges'] * (1 - row['fraction_edges'])).sum() / nrow,
            'auroc': fast_auc(y_true, y_score),
        }
        row['brier'] = row['cal'] + row['ref']
        metrics.append(row)
        del y_score, row
    return metrics

def dataframe_to_summary_metrics(df, metaedge):
    nrow = len(df)
    
    # Compute features and hold in memory as numpy arrays (less memory use than pandas column)
    degree_product = df['source_degree'].values * df['target_degree'].values
    scaled_degree = (degree_product / degree_product.max()).astype(np.float32)
    del degree_product
    
    analytic_prior = xswap.prior.approximate_xswap_prior(df['source_degree'].values,
                                                         df['target_degree'].values,
                                                         df['edge'].sum())
    del df['source_degree'], df['target_degree']
    
    metrics = list()
    for feature in ['xswap_prior', 'scaled_degree', 'analytic_prior']:        
        if feature == 'analytic_prior': 
            df['analytic_prior'] = analytic_prior
            del analytic_prior
        elif feature == 'scaled_degree':
            df['scaled_degree'] = scaled_degree
            del scaled_degree
            
        value_counts = (
            df
            .groupby(["edge", feature])
            .size()
            .rename("count")
            .reset_index()
        )
        auroc = sklearn.metrics.roc_auc_score(value_counts["edge"], value_counts[feature], 
                                              sample_weight=value_counts["count"])
        
        wt = np.where(value_counts["edge"], 1 / (value_counts["edge"] * value_counts["count"]).sum(),
                      1 / (value_counts["count"].sum() - (value_counts["edge"] * value_counts["count"]).sum()))

        weighted_mae = np.average(np.abs(value_counts["edge"] - value_counts[feature]), 
                                  weights=value_counts["count"] * wt)
        del value_counts
        
        # Compute fraction and number of duplicates for each feature value
        row = (
            df
            .groupby(feature)
            .agg({'edge': ['count', 'mean']})
            .reset_index()
        )
        del df[feature]
        
        # Reformat from Dict[Tuple[str, str]: pd.Series] to Dict[str: np.ndarray]
        row = {
            feature: row[(feature, '')].values,
            'count': row[('edge', 'count')].values.astype(np.uint32),
            'fraction_edges': row[('edge', 'mean')].values.astype(np.float32),
        }
        
        # Summarize feature values into single statistics
        row = {
            'feature': feature,
            'metaedge': metaedge,
            'cal': (row['count'] * (row[feature] - row['fraction_edges']) ** 2).sum() / nrow,
            'weighted_mae': weighted_mae,
            'ref': (row['count'] * row['fraction_edges'] * (1 - row['fraction_edges'])).sum() / nrow,
            'auroc': auroc,
        }
        row['brier'] = row['cal'] + row['ref']
        metrics.append(row)
        del row
    return metrics

full_network_all_metrics = list()
sampled_network_all_metrics = list()

full_prior_path = pathlib.Path('../../data/task1/full_priors/')
sampled_prior_path = pathlib.Path('../../data/task1/sampled_priors/')
prior_paths = sorted(list(full_prior_path.glob('*.tsv.gz')))
# prior_paths = [full_prior_path.joinpath("GiG.tsv.gz"), ]
# prior_paths = [full_prior_path.joinpath('AdG.tsv.gz'), ]

prior_paths = list(reversed(prior_paths))

for prior_path in tqdm.tqdm_notebook(prior_paths):
    metaedge = prior_path.name.split('.')[0]
    print(metaedge, flush=True)
    
    # Load DataFrame and produce dataframe with features and outcomes only (save memory)
    prior_df = pd.read_csv(
        prior_path, sep='\t', usecols=['edge', 'source_degree', 'target_degree', 'xswap_prior'],
        dtype={"edge": bool, "source_degree": np.uintc, "target_degree": np.uintc, "xswap_prior": np.single}
    )
    which_nonzero = prior_df.query("(source_degree > 0) & (target_degree > 0)").index
    prior_df = prior_df.loc[which_nonzero]
    
    full_network_all_metrics.extend(dataframe_to_summary_metrics(prior_df, metaedge))
    original_edges = prior_df['edge'].values
    del prior_df
    
    sampled_df = (
        pd.read_csv(f'../../data/task1/sampled_priors/{metaedge}.tsv.gz', sep='\t', 
                    usecols=['edge', 'source_degree', 'target_degree', 'xswap_prior'],
                    dtype={"edge": bool, "source_degree": np.uintc, "target_degree": np.uintc, 
                           "xswap_prior": np.single})
    )
    
    sampled_df = sampled_df.loc[which_nonzero]
    del which_nonzero
    
    sampled_df["original_edges"] = original_edges
    del original_edges
    
    sampled_df = (
        sampled_df
        .query("(source_degree > 0) & (target_degree > 0) & (edge == 0)")
        .drop('edge', axis=1)
        .rename(columns={'original_edges': 'edge'})
    )
    
    sampled_network_all_metrics.extend(dataframe_to_summary_metrics(sampled_df, metaedge))
    del sampled_df['edge'], sampled_df

## Third task: Translating between degree sequences

third_task_all_metrics = list()

full_prior_path = pathlib.Path('../../data/task3/4.data/')
prior_paths = sorted(list(full_prior_path.glob('*.tsv.xz')))
prior_paths = [p for p in prior_paths if p.stem in {"biorxiv.tsv", "ppi.tsv", "tftg.tsv"}]
# prior_paths = [pathlib.Path('full_priors/G<rG.tsv.xz'), ]

for prior_path in tqdm.tqdm_notebook(prior_paths):
    metaedge = prior_path.with_suffix('').stem
    print(metaedge, flush=True)

    prior_df = (
        pd.read_csv(prior_path, sep='\t', usecols=['id_a', 'id_b', 
                                                   'network', 'edge', 'edge_prior'])
        .query('network != "train"')
        .rename(columns={'edge_prior': 'xswap_prior'})
    )

    prior_df = (
        prior_df
        .query('network == "test_recon"')
        .reset_index(drop=True)
        .assign(
            edge_other = prior_df.query('network == "test_new"')['edge'].values,
        )
        .drop(['network'], axis=1)
        .assign(
            source_degree = lambda df: df.groupby('id_a').transform(sum)['edge'],
            target_degree = lambda df: df.groupby('id_b').transform(sum)['edge'],
        )
        .drop(['id_a', 'id_b', 'edge'], axis=1)
        .rename(columns={'edge_other': 'edge'})

    )
    third_task_all_metrics.extend(dataframe_to_summary_metrics(prior_df, metaedge))

full_metrics = pd.DataFrame.from_records(full_network_all_metrics)
sampled_metrics = pd.DataFrame.from_records(sampled_network_all_metrics)
third_task_metrics = pd.DataFrame.from_records(third_task_all_metrics)

all_metrics = pd.concat([
    full_metrics.assign(network='full'),
    sampled_metrics.assign(network='sampled'),
    third_task_metrics.assign(network='other')
])

all_metrics.to_csv('../../data/task1/calibration/hetionet_calibration_metrics.csv', index=False, 
                   columns=['network', 'metaedge', 'feature','cal', 'ref', 'brier', 'weighted_mae', 'auroc'])
