==> nb/1.hetionet_computation/4.compute_feature_rank_correlation.ipynb <==
import pathlib
import regex

import pandas as pd
import scipy.stats
import tqdm

## Full networks

root = pathlib.Path()
prior_paths = sorted(root.glob('*_priors/*.tsv.gz'))

rows = list()

for prior_path in tqdm.tqdm_notebook(prior_paths):
    network = regex.search('.+(?=_priors)', prior_path.parent.name).group()
    metaedge = regex.search('(?<=_priors/).+(?=.tsv.gz)', str(prior_path)).group()
    print(network, metaedge)
    
    df = pd.read_csv(prior_path, sep='\t', usecols=['source_degree', 'target_degree', 'edge', 'xswap_prior'])
    
    xswap_prior = df['xswap_prior'].values
    del df['xswap_prior']

    degree_product = df['source_degree'].values * df['target_degree'].values
    analytic_prior = degree_product / (degree_product - df['source_degree'].values 
                                       - df['target_degree'].values + df['edge'].sum() + 1)
    del df['source_degree'], df['target_degree'], df['edge'], df

    scaled_degree = degree_product / degree_product.max()
    del degree_product

    rows.append({
        'network': network,
        'metaedge': metaedge,
        'xswap_analytic': scipy.stats.spearmanr(xswap_prior, analytic_prior)[0],
        'xswap_degree': scipy.stats.spearmanr(xswap_prior, scaled_degree)[0],
    })
    del xswap_prior, analytic_prior, scaled_degree

corr_df = pd.DataFrame.from_records(rows)
corr_df.to_csv('hetionet_feature_correlation.tsv', sep='\t', index=False)
corr_df.head()
